voice_of_customer/requirements.txt
# Core data processing
numpy>=1.21.0
pandas>=1.3.0
python-docx>=0.8.11
tqdm>=4.62.0

# Text processing
nltk>=3.6.3
spacy>=3.1.0
python-Levenshtein>=0.12.2

# Deep Learning Frameworks
torch>=1.9.0
tensorflow>=2.9.0

# Hugging Face Ecosystem
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
accelerate>=0.20.0

# Keras Ecosystem
keras>=2.12.0
keras-nlp>=0.6.0

# Model Serving
sentencepiece>=0.1.96
protobuf>=3.20.0

# Utilities
python-dotenv>=0.19.0
jupyter>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
